Refactored panel animation from inheritance-based to component-based
Animation has pretty much nothing to do with the behavior of the panel, and so an animating component makes the most sense. The only upside to using inheritance would be for simplicity sake -- it's pretty easy to just have a class inherit. Then again, it's also pretty easy to just add a component.