Added ability to add tabs at runtime (such as map tab)
I refactored TabPanelUI to be able to fill in tab data at runtime. This means that tabs are left blank by default and populated afterwards.

This was added to give us the ability to "unlock" new tabs as the game progresses. Specifically, the map panel should only be made available after the player has found the map (but we may also decide the same for the journal or even the inventory?).